I give up, Huge structural work, mostly on the record-branch.
Turned `Transition` and `State` into records instead, maybe i should do the same to `Machine`? probably not, but i'll keep that in mind.
Removed JavaDoc temporarily.

Removed `StateTransition`, it was a mistake from the get-go.
`ToggleableBoolean` is a workaround (i know) for the immutability of record field references.
`stringToList` and `ListToString` thingies in `Utilities.java`